fix errors arising from refactoring into modules
