+++ src/app/admin/edit-article/edit-article.component.spec.ts
@@ -1,12 +1,12 @@
 import { async, ComponentFixture, TestBed, fakeAsync, tick } from '@angular/core/testing';
 
 import { EditArticleComponent } from './edit-article.component';
-import { AuthenticationService } from '../authentication.service';
+import { AuthenticationService } from '../../authentication.service';
 import { Observable } from 'rxjs/Observable';
 import 'rxjs/add/observable/of';
 import { Article, ArticleType } from './article';
 import { FormsModule } from '@angular/forms';
-import { ArticlesService } from '../articles.service';
+import { ArticlesService } from '../../articles.service';
 import { ActivatedRoute, Router } from '@angular/router';
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 import { MarkdownModule } from 'ngx-md';
