+++ src/app/admin/list-articles/list-articles.component.spec.ts
@@ -1,7 +1,7 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { ListArticlesComponent } from './list-articles.component';
-import { ArticlesService } from '../articles.service';
+import { ArticlesService } from '../../articles.service';
 import { Observable } from 'rxjs/Observable';
 import { RouterTestingModule } from '@angular/router/testing';
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core/';
