+++ src/app/app.component.ts
@@ -3,7 +3,7 @@ import { Component } from '@angular/core';
 import { Observable } from 'rxjs/Observable';
 import * as firebase from 'firebase/app';
 import { ArticlesService } from './articles.service';
-import { Article } from "./edit-article/article";
+import { Article } from './admin/edit-article/article';
 import { AuthenticationService, User } from './authentication.service';
 
 @Component({
